% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addrswm.R
\name{addrswm}
\alias{addrswm}
\title{Add water mark for RStata}
\usage{
addrswm(file, mode = "light", replace = TRUE)
}
\arguments{
\item{file}{image file path}

\item{mode}{light or dark}

\item{replace}{TRUE or FALSE}
}
\description{
Add water mark for RStata
}
