% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addrswm2.R
\name{addrswm2}
\alias{addrswm2}
\title{Add water mark for Geodatasci}
\usage{
addrswm2(file, mode = "light", replace = TRUE)
}
\arguments{
\item{file}{image file path}

\item{mode}{light or dark}

\item{replace}{TRUE or FALSE}
}
\description{
Add water mark for Geodatasci
}
