% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnseg_tokenfun.R
\name{cn_ngrams}
\alias{cn_ngrams}
\title{Chinese text's n-gram token function}
\usage{
cn_ngrams(x, n = 2, n_min = n, ...)
}
\arguments{
\item{x}{just x...}

\item{n}{default is 2}

\item{n_min}{just n_min...}

\item{...}{other parameters}
}
\description{
Chinese text's n-gram token function
}
