% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnseg_tokenfun.R
\name{cnseg_tokenfun}
\alias{cnseg_tokenfun}
\title{Chinese text's segment token function}
\usage{
cnseg_tokenfun(x, ...)
}
\arguments{
\item{x}{just x...}

\item{...}{other parameters}
}
\description{
Chinese text's segment token function
}
