% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cntrade.R
\name{cntrade}
\alias{cntrade}
\title{Download stock data}
\usage{
cntrade(ticker = "", type = "stock")
}
\arguments{
\item{ticker}{Ticker is a stock code or index code to be downloaded from Net. In China, stocks are identified by a six digit numbers, not tickers as in the United States. Examples of codes and the names are as following:
Stock Codes and Stock Names:
  000001 Pingan Bank
  000002 Vank Real Estate Co. Ltd.
  600000 Pudong Development Bank
  600005 Wuhan Steel Co. Ltd.
  900901 INESA Electron Co.,Ltd.
  Index Codes and Index Names:
  000001 The Shanghai Composite Index.
  000300 CSI 300 Index.
  399001 Shenzhen Component Index.
The leading zeros in each code can be omitted.}

\item{type}{Specify that the code is a stock code or index code, 'stock' is the default choice.}
}
\description{
Download historical Market Quotations for a list of stock codes or index codes from Net Ease (a web site providing financial information in China, http://money.163.com/).
}
\examples{
cntrade(1, type = "index")
}
