% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet_tiles.R
\name{gdmap}
\alias{gdmap}
\title{Load GaoDe Map to leaflet}
\usage{
gdmap(map, type = "normal", ...)
}
\arguments{
\item{map}{A leaflet object.}

\item{type}{A character value to set type of Gaode map tiles. Options are "normal" and "satellite".}

\item{...}{Other paramter pass to the addTiles. function}
}
\description{
Simple function like addTiles()
}
\examples{
library(leaflet)
library(leafem)
library(rstatatools)
library(sf)
if(interactive()){
  leaflet() \%>\%
    gdmap(type = "satellite") \%>\%
    addFeatures(locsf, weight = 0.1, radius = 0.1)
}

}
