% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet_tiles.R
\name{gdmap_annotion}
\alias{gdmap_annotion}
\title{Load GaoDe Map annotion to leaflet}
\usage{
gdmap_annotion(map, ...)
}
\arguments{
\item{map}{A leaflet object.}

\item{...}{Other paramter pass to the addTiles. function}
}
\description{
Simple function like addTiles()
}
\examples{
library(leaflet)
library(leafem)
library(rstatatools)
library(sf)
if(interactive()){
  leaflet() \%>\%
    gdmap(type = "satellite") \%>\%
    gdmap_annotion() \%>\%
    addFeatures(locsf, weight = 0.1, radius = 0.1)
}

}
