% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet_tiles.R
\name{geoqmap}
\alias{geoqmap}
\title{Load Geoq Map to leaflet}
\usage{
geoqmap(map, type = "normal", ...)
}
\arguments{
\item{map}{A leaflet object.}

\item{type}{A character value to set type of Geoq map tiles. Options are "normal", "PurplishBlue", "Gray", "Warm", "ENG", "LabelAndBoundaryLine", "Subway", "WorldHydroMap", "Gray_OnlySymbol", "Gray_Reference", "PurplishBlue_OnlySymbol", "PurplishBlue_Reference", "Warm_OnlySymbol", "Warm_Reference".}

\item{...}{Other paramter pass to the addTiles. function}
}
\description{
Simple function like addTiles()
}
\examples{
library(leaflet)
library(leafem)
library(rstatatools)
library(sf)
if(interactive()){
  leaflet() \%>\%
    geoqmap(type = "ENG") \%>\%
    addFeatures(locsf, weight = 0.1, radius = 0.1)
}

}
