% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coords.R
\name{get_coords}
\alias{get_coords}
\title{Geocode with Gaodemap(amap)}
\usage{
get_coords(address, city = NULL, key = Sys.getenv("amap.key"))
}
\arguments{
\item{address}{just address}

\item{city}{city, default if NULL}

\item{key}{amap key, Sys.getenv("amap.key") as default.}
}
\description{
Geocode with Gaodemap(amap)
}
