% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coords.R
\name{get_coords_bd}
\alias{get_coords_bd}
\title{Geocode with Baidumap}
\usage{
get_coords_bd(address, city = NULL, key = Sys.getenv("baidumap.key"))
}
\arguments{
\item{address}{just address}

\item{city}{city, default if NULL}

\item{key}{Baidu key, Sys.getenv("baidumap.key") as default.}
}
\description{
Geocode with Baidumap
}
