% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet_tiles.R
\name{tdtmap}
\alias{tdtmap}
\title{Load Tiandi Map to leaflet}
\usage{
tdtmap(map, type = "normal", ...)
}
\arguments{
\item{map}{A leaflet object.}

\item{type}{A character value to set type of Tiandi map tiles. Options are "normal", "satellite", "terrain".}

\item{...}{Other paramter pass to the addTiles. function}
}
\description{
Simple function like addTiles()
}
\examples{
library(leaflet)
library(leafem)
library(rstatatools)
library(sf)
if(interactive()){
  leaflet() \%>\%
    tdtmap(type = "terrain") \%>\%
    addFeatures(locsf, weight = 0.1, radius = 0.1)
}
}
